pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["local"]


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.74.0"
uvicorn = "^0.17.5"
pendulum = "^2.1.2"
PyJWT = "^2.3.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
tomlkit = "^0.10.0"
psycopg2-binary = "^2.9.3"
pony = "^0.7.16"
httpx = "^0.22.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
